+++ packages/railtracks/src/railtracks/__init__.py
@@ -42,7 +42,6 @@ __all__ = [
     "evaluations",
     "vector_stores",
     "rag",
-    "RagConfig",
     "Flow",
     "enable_logging",
     "wrap_node",
@@ -53,10 +52,10 @@ __all__ = [
     "wrap_llm",
 ]
 
-from railtracks.built_nodes.concrete.rag import RagConfig
-from railtracks.built_nodes.easy_usage_wrappers import (
-    agent_node,
+
+from railtracks.built_nodes import (
     function_node,
+    agent_node,
 )
 
 from . import (
