+++ packages/railtracks/src/railtracks/built_nodes/function/__init__.py
@@ -1,0 +1,6 @@
+__all__ = [
+    "function_node",
+]
+
+
+from .node import function_node
