+++ packages/railtracks/src/railtracks/built_nodes/function/node.py
@@ -18,7 +18,7 @@ from typing import (
 )
 
 from railtracks.built_nodes._node_builder import NodeBuilder
-from railtracks.built_nodes.concrete.function_base import RTFunction
+from railtracks.built_nodes.function.base import RTFunction
 from railtracks.exceptions import NodeCreationError
 from railtracks.middleware.core import Middleware
 from railtracks.nodes.manifest import ToolManifest
@@ -28,46 +28,10 @@ from railtracks.validation.node_creation.validation import (
     validate_tool_manifest_against_function,
 )
 
+from .base import CallableAsyncRTFunction, CallableSyncRTFunction
+
 _TOutput = TypeVar("_TOutput")
 _P = ParamSpec("_P")
-
-
-class CallableSyncRTFunction(RTFunction[_P, _TOutput], Generic[_P, _TOutput]):
-    def __init__(
-        self, func: Callable[_P, _TOutput], node_type: type[Node[_P, _TOutput]]
-    ):
-        self.func = func
-        self.node_type = node_type
-        functools.update_wrapper(self, func, updated=())
-
-    def __call__(self, *args: _P.args, **kwargs: _P.kwargs) -> _TOutput:
-        return self.func(*args, **kwargs)
-
-    def with_node_type(
-        self, node_type: type[Node[_P, _TOutput]]
-    ) -> CallableSyncRTFunction[_P, _TOutput]:
-        """Returns a copy of this CallableSyncRTFunction with a different `node_type`. Does not modify this instance."""
-        return CallableSyncRTFunction(self.func, node_type)
-
-
-class CallableAsyncRTFunction(RTFunction[_P, _TOutput], Generic[_P, _TOutput]):
-    def __init__(
-        self,
-        func: Callable[_P, Coroutine[None, None, _TOutput]],
-        node_type: type[Node[_P, _TOutput]],
-    ):
-        self.func = func
-        self.node_type = node_type
-        functools.update_wrapper(self, func, updated=())
-
-    async def __call__(self, *args: _P.args, **kwargs: _P.kwargs) -> _TOutput:
-        return await self.func(*args, **kwargs)
-
-    def with_node_type(
-        self, node_type: type[Node[_P, _TOutput]]
-    ) -> CallableAsyncRTFunction[_P, _TOutput]:
-        """Returns a copy of this CallableAsyncRTFunction with a different `node_type`. Does not modify this instance."""
-        return CallableAsyncRTFunction(self.func, node_type)
 
 
 # note there is an intentional overlap in overloads
