+++ packages/railtracks/src/railtracks/built_nodes/llm/llm_helpers.py
@@ -8,7 +8,7 @@ from typing import (
 )
 
 from pydantic import BaseModel
-from railtracks.built_nodes.concrete.response import StringResponse, StructuredResponse
+from railtracks.built_nodes.llm.response import StringResponse, StructuredResponse
 from railtracks.built_nodes.llm.model_invoker import ModelInvoker
 from railtracks.exceptions.errors import LLMError, NodeInvocationError
 from railtracks.interaction._call import call
