+++ packages/railtracks/src/railtracks/built_nodes/llm/node.py
@@ -3,10 +3,10 @@ from typing import Iterable, Literal, Type, TypeVar, overload
 from pydantic import BaseModel
 
 from railtracks.built_nodes._types import ModelSource
-from railtracks.built_nodes.concrete import (
+from railtracks.built_nodes.function.base import (
     RTFunction,
 )
-from railtracks.built_nodes.concrete.response import StringResponse, StructuredResponse
+from railtracks.built_nodes.llm.response import StringResponse, StructuredResponse
 from railtracks.built_nodes.llm.middleware.core import ModelMiddleware
 from railtracks.llm.message import SystemMessage
 from railtracks.llm.tools.parameters._base import Parameter
