+++ packages/railtracks/tests/integration_tests/guardrails/test_agent_node_guardrails.py
@@ -6,7 +6,7 @@ import pytest
 import railtracks as rt
 from pydantic import BaseModel, Field
 
-from railtracks.built_nodes.concrete.response import StringResponse, StructuredResponse
+from railtracks.built_nodes.llm.response import StringResponse, StructuredResponse
 from railtracks.guardrails import GuardrailBlockedError, GuardrailDecision, InputGuard, LLMGuardrailEvent, OutputGuard
 from railtracks.llm import AssistantMessage
