+++ packages/railtracks/tests/integration_tests/guardrails/test_guardrail_fail_open.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 import pytest
 import railtracks as rt
 
-from railtracks.built_nodes.concrete.response import StringResponse
+from railtracks.built_nodes.llm.response import StringResponse
 from railtracks.guardrails.core import GuardrailBlockedError, GuardrailDecision, InputGuard, LLMGuardrailEvent, OutputGuard
