+++ packages/railtracks/tests/integration_tests/library/test_tool_calling_llms.py
@@ -1,6 +1,6 @@
 import pytest
 import railtracks as rt
-from railtracks.built_nodes.concrete.response import StringResponse
+from railtracks.built_nodes.llm.response import StringResponse
 from railtracks.exceptions import NodeCreationError
 from railtracks.llm import AssistantMessage, ToolCall
 from railtracks.llm.response import Response
